+++ CustomBL/Services/ICustomService.cs
@@ -5,8 +5,8 @@ namespace Custom.BL.Services
 {
     public interface ICustomService
     {
-        double GetCarCustomValue(FuelType fuelType, int engineVolume,  int price = default, DateTime year = default);
-        double GetTruckCustomValue(int price, DateTime year, int engineVolume, int fullWeight);
-        double GetBikeCustomValue(int price, DateTime year, int engineVolume);
+        int GetCarCustomValue(FuelType fuelType, int engineVolume, int price = default, DateTime year = default);
+        int GetTruckCustomValue(int price, DateTime year, int engineVolume, int fullWeight);
+        int GetBikeCustomValue(int price, DateTime year, int engineVolume);
     }
 }
